Remove sample Id from query to get viral insertion choice

+++ patient-reporter/src/main/java/com/hartwig/hmftools/patientreporter/copynumber/CopyNumberAnalysis.java
@@ -36,7 +36,7 @@ public abstract class CopyNumberAnalysis {
 
     public abstract double tumorMutationalBurdenPerMb();
 
-    public abstract double tumorMutationalLoad();
+    public abstract int tumorMutationalLoad();
 
     @NotNull
     public abstract MicrosatelliteStatus microsatelliteStatus();
